split distributed coords tests
