+++ tests/e2e/high_availability/CMakeLists.txt
@@ -3,7 +3,15 @@ find_package(gflags REQUIRED)
 copy_e2e_python_files(high_availability coordinator.py)
 copy_e2e_python_files(high_availability single_coordinator.py)
 copy_e2e_python_files(high_availability coord_cluster_registration.py)
+copy_e2e_python_files(high_availability distributed_coords_common.py)
 copy_e2e_python_files(high_availability distributed_coords.py)
+copy_e2e_python_files(high_availability distributed_coords_part_2.py)
+copy_e2e_python_files(high_availability distributed_coords_part_3.py)
+copy_e2e_python_files(high_availability distributed_coords_part_4.py)
+copy_e2e_python_files(high_availability distributed_coords_part_5.py)
+copy_e2e_python_files(high_availability distributed_coords_part_6.py)
+copy_e2e_python_files(high_availability distributed_coords_part_7.py)
+copy_e2e_python_files(high_availability distributed_coords_part_8.py)
 copy_e2e_python_files(high_availability disable_writing_on_main_after_restart.py)
 copy_e2e_python_files(high_availability manual_setting_replicas.py)
 copy_e2e_python_files(high_availability not_replicate_from_old_main.py)
