+++ tests/e2e/high_availability/distributed_coords_part_8.py
@@ -1,0 +1,359 @@
+# Copyright 2024 Memgraph Ltd.
+#
+# Use of this software is governed by the Business Source License
+# included in the file licenses/BSL.txt; by using this file, you agree to be bound by the terms of the Business Source
+# License, and you may not use this file except in compliance with the Business Source License.
+#
+# As of the Change Date specified in that file, in accordance with
+# the Business Source License, use of this software will be governed
+# by the Apache License, Version 2.0, included in the file
+# licenses/APL.txt.
+
+import concurrent.futures
+import os
+import sys
+import tempfile
+
+import interactive_mg_runner
+import pytest
+from common import (
+    connect,
+    execute_and_fetch_all,
+    find_instance_and_assert_instances,
+    ignore_elapsed_time_from_results,
+    update_tuple_value,
+)
+from distributed_coords_common import (
+    get_default_setup_queries,
+    get_instances_description_no_setup,
+)
+from mg_utils import (
+    mg_assert_until,
+    mg_sleep_and_assert,
+    mg_sleep_and_assert_collection,
+    mg_sleep_and_assert_multiple,
+    wait_for_status_change,
+)
+
+interactive_mg_runner.SCRIPT_DIR = os.path.dirname(os.path.realpath(__file__))
+interactive_mg_runner.PROJECT_DIR = os.path.normpath(
+    os.path.join(interactive_mg_runner.SCRIPT_DIR, "..", "..", "..", "..")
+)
+interactive_mg_runner.BUILD_DIR = os.path.normpath(os.path.join(interactive_mg_runner.PROJECT_DIR, "build"))
+interactive_mg_runner.MEMGRAPH_BINARY = os.path.normpath(os.path.join(interactive_mg_runner.BUILD_DIR, "memgraph"))
+
+
+def test_all_coords_down_resume():
+    # Goal of this test is to check that coordinators are able to resume, if cluster is able to resume where it
+    # left off if all coordinators go down
+
+    # 1. Start cluster
+    # 2. Check everything works correctly
+    # 3. Stop all coordinators
+    # 4. Start 2 coordinators, one should be leader, and one follower
+    # 5. Check everything works correctly
+
+    # 1
+    temp_dir = tempfile.TemporaryDirectory()
+    temp_dir_name = temp_dir.name
+
+    inner_instances_description = get_instances_description_no_setup(
+        temp_dir_name, test_name="test_all_coords_down_resume", use_durability=True
+    )
+
+    interactive_mg_runner.start_all(inner_instances_description, keep_directories=False)
+
+    coord_cursor_3 = connect(host="localhost", port=7692).cursor()
+    for query in get_default_setup_queries():
+        execute_and_fetch_all(coord_cursor_3, query)
+
+    # 2
+    def show_instances_coord3():
+        return ignore_elapsed_time_from_results(sorted(list(execute_and_fetch_all(coord_cursor_3, "SHOW INSTANCES;"))))
+
+    coord_cursor_1 = connect(host="localhost", port=7690).cursor()
+
+    def show_instances_coord1():
+        coord_cursor_1 = connect(host="localhost", port=7690).cursor()
+        return ignore_elapsed_time_from_results(sorted(list(execute_and_fetch_all(coord_cursor_1, "SHOW INSTANCES;"))))
+
+    coord_cursor_2 = connect(host="localhost", port=7691).cursor()
+
+    def show_instances_coord2():
+        return ignore_elapsed_time_from_results(sorted(list(execute_and_fetch_all(coord_cursor_2, "SHOW INSTANCES;"))))
+
+    data = [
+        ("coordinator_1", "localhost:7690", "localhost:10111", "localhost:10121", "up", "follower"),
+        ("coordinator_2", "localhost:7691", "localhost:10112", "localhost:10122", "up", "follower"),
+        ("coordinator_3", "localhost:7692", "localhost:10113", "localhost:10123", "up", "leader"),
+        ("instance_1", "localhost:7687", "", "localhost:10011", "up", "replica"),
+        ("instance_2", "localhost:7688", "", "localhost:10012", "up", "replica"),
+        ("instance_3", "localhost:7689", "", "localhost:10013", "up", "main"),
+    ]
+
+    mg_sleep_and_assert(data, show_instances_coord3)
+    mg_sleep_and_assert(data, show_instances_coord1)
+    mg_sleep_and_assert(data, show_instances_coord2)
+
+    # 3
+
+    interactive_mg_runner.kill(inner_instances_description, "coordinator_1")
+    interactive_mg_runner.kill(inner_instances_description, "coordinator_2")
+    interactive_mg_runner.kill(inner_instances_description, "coordinator_3")
+
+    # 4
+
+    with concurrent.futures.ThreadPoolExecutor(2) as executor:
+        executor.submit(interactive_mg_runner.start, inner_instances_description, "coordinator_2")
+        executor.submit(interactive_mg_runner.start, inner_instances_description, "coordinator_1")
+
+    # 5
+
+    leader_data = [
+        ("coordinator_1", "localhost:7690", "localhost:10111", "localhost:10121", "up", "follower"),
+        ("coordinator_2", "localhost:7691", "localhost:10112", "localhost:10122", "up", "follower"),
+        ("coordinator_3", "localhost:7692", "localhost:10113", "localhost:10123", "down", "follower"),
+        ("instance_1", "localhost:7687", "", "localhost:10011", "up", "replica"),
+        ("instance_2", "localhost:7688", "", "localhost:10012", "up", "replica"),
+        ("instance_3", "localhost:7689", "", "localhost:10013", "up", "replica"),
+    ]
+
+    wait_for_status_change(show_instances_coord1, {"coordinator_1", "coordinator_2"}, "leader")
+
+    leader = find_instance_and_assert_instances(
+        instance_role="leader", num_coordinators=3, coord_ids_to_skip_validation={3}
+    )
+
+    main = find_instance_and_assert_instances(
+        instance_role="main", num_coordinators=3, coord_ids_to_skip_validation={3}
+    )
+    leader_data = update_tuple_value(leader_data, leader, 0, -1, "leader")
+    leader_data = update_tuple_value(leader_data, main, 0, -1, "main")
+
+    coord_cursor_1 = connect(host="localhost", port=7690).cursor()
+    coord_cursor_2 = connect(host="localhost", port=7691).cursor()
+
+    mg_sleep_and_assert(leader_data, show_instances_coord1)
+
+    mg_sleep_and_assert(leader_data, show_instances_coord2)
+
+    # 6
+    interactive_mg_runner.kill(inner_instances_description, "instance_3")
+    interactive_mg_runner.start(inner_instances_description, "coordinator_3")
+
+    # 7
+
+    leader_data = [
+        ("coordinator_1", "localhost:7690", "localhost:10111", "localhost:10121", "up", "follower"),
+        ("coordinator_2", "localhost:7691", "localhost:10112", "localhost:10122", "up", "follower"),
+        ("coordinator_3", "localhost:7692", "localhost:10113", "localhost:10123", "up", "follower"),
+        ("instance_1", "localhost:7687", "", "localhost:10011", "up", "replica"),
+        ("instance_2", "localhost:7688", "", "localhost:10012", "up", "replica"),
+        ("instance_3", "localhost:7689", "", "localhost:10013", "down", "unknown"),
+    ]
+
+    wait_for_status_change(show_instances_coord1, {"instance_3"}, "unknown")
+
+    leader = find_instance_and_assert_instances(instance_role="leader", num_coordinators=3)
+
+    main = find_instance_and_assert_instances(instance_role="main", num_coordinators=3)
+
+    leader_data = update_tuple_value(leader_data, leader, 0, -1, "leader")
+    leader_data = update_tuple_value(leader_data, main, 0, -1, "main")
+
+    coord_cursor_3 = connect(host="localhost", port=7692).cursor()
+
+    mg_sleep_and_assert(leader_data, show_instances_coord1)
+
+    mg_sleep_and_assert(leader_data, show_instances_coord2)
+
+    mg_sleep_and_assert(leader_data, show_instances_coord3)
+
+    interactive_mg_runner.stop_all(keep_directories=False)
+
+
+def test_one_coord_down_with_durability_resume():
+    # Goal of this test is to check that coordinators are able to resume, if cluster is able to resume where it
+    # left off if all coordinators go down
+
+    # 1. Start cluster
+    # 2. Check everything works correctly
+    # 3. Stop all coordinators
+    # 4. Start 2 coordinators, one should be leader, and one follower
+    # 5. Check everything works correctly
+
+    # 1
+    temp_dir = tempfile.TemporaryDirectory()
+    temp_dir_name = temp_dir.name
+
+    inner_instances_description = get_instances_description_no_setup(
+        temp_dir_name, test_name="test_one_coord_down_with_durability_resume", use_durability=True
+    )
+
+    interactive_mg_runner.start_all(inner_instances_description, keep_directories=False)
+
+    coord_cursor_3 = connect(host="localhost", port=7692).cursor()
+
+    for query in get_default_setup_queries():
+        execute_and_fetch_all(coord_cursor_3, query)
+
+    # 2
+    def show_instances_coord3():
+        return ignore_elapsed_time_from_results(sorted(list(execute_and_fetch_all(coord_cursor_3, "SHOW INSTANCES;"))))
+
+    coord_cursor_1 = connect(host="localhost", port=7690).cursor()
+
+    def show_instances_coord1():
+        return ignore_elapsed_time_from_results(sorted(list(execute_and_fetch_all(coord_cursor_1, "SHOW INSTANCES;"))))
+
+    coord_cursor_2 = connect(host="localhost", port=7691).cursor()
+
+    def show_instances_coord2():
+        return ignore_elapsed_time_from_results(sorted(list(execute_and_fetch_all(coord_cursor_2, "SHOW INSTANCES;"))))
+
+    data = [
+        ("coordinator_1", "localhost:7690", "localhost:10111", "localhost:10121", "up", "follower"),
+        ("coordinator_2", "localhost:7691", "localhost:10112", "localhost:10122", "up", "follower"),
+        ("coordinator_3", "localhost:7692", "localhost:10113", "localhost:10123", "up", "leader"),
+        ("instance_1", "localhost:7687", "", "localhost:10011", "up", "replica"),
+        ("instance_2", "localhost:7688", "", "localhost:10012", "up", "replica"),
+        ("instance_3", "localhost:7689", "", "localhost:10013", "up", "main"),
+    ]
+
+    mg_sleep_and_assert(data, show_instances_coord3)
+    mg_sleep_and_assert(data, show_instances_coord1)
+    mg_sleep_and_assert(data, show_instances_coord2)
+
+    # 3
+
+    interactive_mg_runner.kill(inner_instances_description, "coordinator_1")
+
+    leader_data = [
+        ("coordinator_1", "localhost:7690", "localhost:10111", "localhost:10121", "down", "follower"),
+        ("coordinator_2", "localhost:7691", "localhost:10112", "localhost:10122", "up", "follower"),
+        ("coordinator_3", "localhost:7692", "localhost:10113", "localhost:10123", "up", "leader"),
+        ("instance_1", "localhost:7687", "", "localhost:10011", "up", "replica"),
+        ("instance_2", "localhost:7688", "", "localhost:10012", "up", "replica"),
+        ("instance_3", "localhost:7689", "", "localhost:10013", "up", "main"),
+    ]
+    mg_sleep_and_assert(leader_data, show_instances_coord3)
+
+    # 4
+    interactive_mg_runner.start(inner_instances_description, "coordinator_1")
+
+    # 5
+
+    data = [
+        ("coordinator_1", "localhost:7690", "localhost:10111", "localhost:10121", "up", "follower"),
+        ("coordinator_2", "localhost:7691", "localhost:10112", "localhost:10122", "up", "follower"),
+        ("coordinator_3", "localhost:7692", "localhost:10113", "localhost:10123", "up", "leader"),
+        ("instance_1", "localhost:7687", "", "localhost:10011", "up", "replica"),
+        ("instance_2", "localhost:7688", "", "localhost:10012", "up", "replica"),
+        ("instance_3", "localhost:7689", "", "localhost:10013", "up", "main"),
+    ]
+
+    coord_cursor_1 = connect(host="localhost", port=7690).cursor()
+
+    mg_sleep_and_assert(data, show_instances_coord3)
+    mg_sleep_and_assert(data, show_instances_coord1)
+    mg_sleep_and_assert(data, show_instances_coord2)
+
+    # 6
+    interactive_mg_runner.kill(inner_instances_description, "instance_3")
+
+    # 7
+
+    leader_data = [
+        ("coordinator_1", "localhost:7690", "localhost:10111", "localhost:10121", "up", "follower"),
+        ("coordinator_2", "localhost:7691", "localhost:10112", "localhost:10122", "up", "follower"),
+        ("coordinator_3", "localhost:7692", "localhost:10113", "localhost:10123", "up", "leader"),
+        ("instance_1", "localhost:7687", "", "localhost:10011", "up", "main"),
+        ("instance_2", "localhost:7688", "", "localhost:10012", "up", "replica"),
+        ("instance_3", "localhost:7689", "", "localhost:10013", "down", "unknown"),
+    ]
+
+    mg_sleep_and_assert(leader_data, show_instances_coord3)
+    mg_sleep_and_assert(leader_data, show_instances_coord1)
+    mg_sleep_and_assert(leader_data, show_instances_coord2)
+
+    interactive_mg_runner.stop_all(keep_directories=False)
+
+
+def test_registration_does_not_deadlock_when_instance_is_down():
+    # Goal of this test is to assert that system doesn't deadlock in case of failure on registration
+
+    # 1
+    temp_dir = tempfile.TemporaryDirectory()
+    temp_dir_name = temp_dir.name
+
+    interactive_mg_runner.stop_all(keep_directories=False)
+    inner_instances_description = get_instances_description_no_setup(
+        temp_dir_name, test_name="test_registration_does_not_deadlock_when_instance_is_down", use_durability=True
+    )
+
+    interactive_mg_runner.start(inner_instances_description, "coordinator_1")
+    interactive_mg_runner.start(inner_instances_description, "coordinator_2")
+    interactive_mg_runner.start(inner_instances_description, "coordinator_3")
+
+    setup_queries = [
+        "ADD COORDINATOR 1 WITH CONFIG {'bolt_server': 'localhost:7690', 'coordinator_server': 'localhost:10111', 'management_server': 'localhost:10121'}",
+        "ADD COORDINATOR 2 WITH CONFIG {'bolt_server': 'localhost:7691', 'coordinator_server': 'localhost:10112', 'management_server': 'localhost:10122'}",
+    ]
+
+    coord_cursor_3 = connect(host="localhost", port=7692).cursor()
+    for query in setup_queries:
+        execute_and_fetch_all(coord_cursor_3, query)
+
+    query = "REGISTER INSTANCE instance_1 WITH CONFIG {'bolt_server': 'localhost:7687', 'management_server': 'localhost:10011', 'replication_server': 'localhost:10001'};"
+
+    with pytest.raises(Exception) as e:
+        execute_and_fetch_all(coord_cursor_3, query)
+
+    assert "Couldn't register replica instance because setting instance to replica failed!" in str(e.value)
+
+    interactive_mg_runner.start(inner_instances_description, "instance_1")
+    execute_and_fetch_all(coord_cursor_3, query)
+
+    interactive_mg_runner.start(inner_instances_description, "instance_2")
+    interactive_mg_runner.start(inner_instances_description, "instance_3")
+
+    setup_queries = [
+        "REGISTER INSTANCE instance_2 WITH CONFIG {'bolt_server': 'localhost:7688', 'management_server': 'localhost:10012', 'replication_server': 'localhost:10002'};",
+        "REGISTER INSTANCE instance_3 WITH CONFIG {'bolt_server': 'localhost:7689', 'management_server': 'localhost:10013', 'replication_server': 'localhost:10003'};",
+        "SET INSTANCE instance_3 TO MAIN",
+    ]
+
+    for query in setup_queries:
+        execute_and_fetch_all(coord_cursor_3, query)
+
+    def show_instances_coord3():
+        return ignore_elapsed_time_from_results(sorted(list(execute_and_fetch_all(coord_cursor_3, "SHOW INSTANCES;"))))
+
+    coord_cursor_1 = connect(host="localhost", port=7690).cursor()
+
+    def show_instances_coord1():
+        return ignore_elapsed_time_from_results(sorted(list(execute_and_fetch_all(coord_cursor_1, "SHOW INSTANCES;"))))
+
+    coord_cursor_2 = connect(host="localhost", port=7691).cursor()
+
+    def show_instances_coord2():
+        return ignore_elapsed_time_from_results(sorted(list(execute_and_fetch_all(coord_cursor_2, "SHOW INSTANCES;"))))
+
+    data = [
+        ("coordinator_1", "localhost:7690", "localhost:10111", "localhost:10121", "up", "follower"),
+        ("coordinator_2", "localhost:7691", "localhost:10112", "localhost:10122", "up", "follower"),
+        ("coordinator_3", "localhost:7692", "localhost:10113", "localhost:10123", "up", "leader"),
+        ("instance_1", "localhost:7687", "", "localhost:10011", "up", "replica"),
+        ("instance_2", "localhost:7688", "", "localhost:10012", "up", "replica"),
+        ("instance_3", "localhost:7689", "", "localhost:10013", "up", "main"),
+    ]
+
+    mg_sleep_and_assert(data, show_instances_coord3)
+    mg_sleep_and_assert(data, show_instances_coord1)
+    mg_sleep_and_assert(data, show_instances_coord2)
+
+    interactive_mg_runner.stop_all(keep_directories=False)
+
+
+if __name__ == "__main__":
+    sys.exit(pytest.main([__file__, "-rA"]))
